refactor: FilteringCell 내부 변수 CategoryID타입으로 변경
String타입에서 Categoryid타입으로 변경
